fixing output issue

+++ print_an_integer.c
@@ -36,7 +36,7 @@ int print_recursive(int num)
  */
 int print_an_integer(const char *format, va_list args_list)
 {
-	int num = va_arg(args_list, int);
+	long int num = va_arg(args_list, long int);
 	int sum = 0;
 
 	if (*format == 'l' || *format == 'h')
